requirements.txt
astrapy==1.4.1
cassandra-driver==3.29.2
email_validator==2.2.0
fastapi[all]
uvicorn
requests
pydantic
fastapi-cli==0.0.5
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
numpy==2.1.1
openai==1.43.0
pydantic_core==2.23.4
python-dotenv==1.0.1
websockets==13.1